+++ src/test/java/seedu/address/report/ReportGeneratorTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
+import seedu.address.storage.ReportGenerator;
 
 //@@author bernicechio
 public class ReportGeneratorTest {
